extract roles section
