+++ components/list/ListItemWithIcon.tsx
@@ -1,4 +1,4 @@
-import { useTheme } from '@mui/material'
+import { Box, Stack, useTheme } from '@mui/material'
 import ListItem from '@mui/material/ListItem'
 import ListItemIcon from '@mui/material/ListItemIcon'
 import ListItemText from '@mui/material/ListItemText'
@@ -9,12 +9,14 @@ type ListItemWithIconProps = {
   listIcon?: ReactNode
   listText?: string
   sxProps?: object
+  children?: ReactNode
 }
 
 const ListItemWithIcon = ({
   listIcon,
   listText,
   sxProps,
+  children
 }: ListItemWithIconProps) => {
   const theme = useTheme()
   const palette = theme.palette
@@ -22,7 +24,6 @@ const ListItemWithIcon = ({
   return (
     <ListItem
       sx={{
-        height: '3.5rem',
         overflowWrap: 'break-word',
         boxShadow:
           '0px 4px 8px 2px rgba(52, 61, 62, 0.04), 0px 2px 4px rgba(52, 61, 62, 0.04)',
@@ -37,16 +38,19 @@ const ListItemWithIcon = ({
       {listText && (
         <ListItemText
           primary={
-            <Typography
-              variant="labelMedium"
-              color={palette.secondary.contrastText}
-              sx={{
-                textTransform: 'capitalize',
-                display: 'inline-block',
-              }}
-            >
-              {listText}
-            </Typography>
+            <Stack spacing={'1rem'}>
+              <Typography
+                variant="labelMedium"
+                color={palette.secondary.contrastText}
+                sx={{
+                  textTransform: 'capitalize',
+                  display: 'inline-block',
+                }}
+              >
+                {listText}
+              </Typography>
+              {children}
+            </Stack>
           }
         />
       )}
